Remove finished song before fadeout
Previously, we had a problem when there was only one song in a playlist. When it was finished, the fadeout was triggered. At the same time, we tried to start playing it, but the PlaybackService found the song still active and tried to continue the already finished song.
By removing the finished song before fadeout, we can be certain that it is removed and the Play logic works as expected.