Only wait for song end when it is not finished
Previously, when pausing a song just before fadeout would be triggered, we had a problem when resuming that song. E.g., it would only play a remaining second, but our expected song end (including the fadeout time) would then be -4 secs. With a negative wait time, the SongEnded would then never be triggered.